server/requirements.txt
Flask
openai
python-dotenv
sounddevice
wavio
numpy
